move `fulcio` utils out of `pkg`
