+++ cmd/cosign/cli/fulcio/fulcio.go
